+++ extensions/cua-computer/src/commands.ts
@@ -33,7 +33,7 @@ import {
 } from "./frame.js";
 import { createCuaMcpDriver } from "./mcp-driver-client.js";
 import { closeRecordingExecution } from "./recording-actions.js";
-import { handleV2Act, type CuaComputerActParams } from "./v2-actions.js";
+import { handleWindowAct, type CuaComputerActParams } from "./window-actions.js";
 
 const AVAILABILITY_POLL_MS = 5_000;
 const CUA_WIRE_ACTION_NAMES = COMPUTER_USE_V2_ACTION_NAMES.slice(1, 14);
@@ -278,37 +278,37 @@ async function handleDesktopAct(
   if (!(CUA_WIRE_ACTION_NAMES as readonly string[]).includes(params.action)) {
     throw new Error(`COMPUTER_UNSUPPORTED_ACTION: ${params.action}`);
   }
-  const v1Params = params as CuaComputerActParams;
-  assertPrimaryDisplay(v1Params.screenIndex);
+  const desktopParams = params as CuaComputerActParams;
+  assertPrimaryDisplay(desktopParams.screenIndex);
   // `wait` never reaches the wire: core sleeps locally and the Swift wire enum
   // has no wait case, so accepting it here would fork the computer.act contract.
   if (
-    v1Params.action === "hold_key" ||
-    v1Params.action === "left_mouse_down" ||
-    v1Params.action === "left_mouse_up"
+    desktopParams.action === "hold_key" ||
+    desktopParams.action === "left_mouse_down" ||
+    desktopParams.action === "left_mouse_up"
   ) {
     // Upstream has no desktop keyboard-down API, and its Linux mouse hold tools
     // are window-only, so these actions cannot preserve desktop-scope semantics.
-    throw new Error(`COMPUTER_UNSUPPORTED_ACTION: ${v1Params.action}`);
+    throw new Error(`COMPUTER_UNSUPPORTED_ACTION: ${desktopParams.action}`);
   }
 
   // Every action uses scope:"desktop", a global SendInput/XTest/wayland_desktop
   // injection that is inherently foreground and ignores delivery_mode (that
   // background-vs-foreground contract is window-targeted only). We deliberately
   // never send delivery_mode.
-  switch (v1Params.action) {
+  switch (desktopParams.action) {
     case "type": {
-      if (!v1Params.text) {
+      if (!desktopParams.text) {
         throw new Error("COMPUTER_INVALID_REQUEST: text is required for type");
       }
-      assertToolSuccess(await driver.typeText(v1Params.text, signal), "type_text");
+      assertToolSuccess(await driver.typeText(desktopParams.text, signal), "type_text");
       break;
     }
     case "key": {
       // press_key applies the modifier array on every backend: X11 via XTest,
       // and native Wayland by internally promoting a modifier chord to
       // hotkey_focused. No separate hotkey call is needed for chords.
-      const chord = parseKeyChord(v1Params.keys);
+      const chord = parseKeyChord(desktopParams.keys);
       assertToolSuccess(
         await driver.pressKey(
           {
@@ -322,10 +322,10 @@ async function handleDesktopAct(
       break;
     }
     case "scroll": {
-      if (!v1Params.scrollDirection) {
+      if (!desktopParams.scrollDirection) {
         throw new Error("COMPUTER_INVALID_REQUEST: scrollDirection is required for scroll");
       }
-      if (normalizeModifiers(v1Params.modifiers).length > 0) {
+      if (normalizeModifiers(desktopParams.modifiers).length > 0) {
         throw new Error(
           "COMPUTER_UNSUPPORTED_ACTION: modifier-held scroll is unsupported by cua-driver",
         );
@@ -334,20 +334,20 @@ async function handleDesktopAct(
       // frame-authorized like clicks. We deliberately do not synthesize a point
       // from get_cursor_position: that mixes cursor and capture coordinate
       // spaces across X11/Wayland/Windows and would scroll an unverified target.
-      const frame = await currentFrame(driver, frameState, v1Params, signal);
-      const point = scalePoint(frame, v1Params.x, v1Params.y, v1Params.action);
+      const frame = await currentFrame(driver, frameState, desktopParams, signal);
+      const point = scalePoint(frame, desktopParams.x, desktopParams.y, desktopParams.action);
       const direction = {
         up: ScrollDirection.Up,
         down: ScrollDirection.Down,
         left: ScrollDirection.Left,
         right: ScrollDirection.Right,
-      }[v1Params.scrollDirection];
+      }[desktopParams.scrollDirection];
       assertToolSuccess(
         await driver.scroll(
           {
             direction,
             // Schema guarantees a positive amount; cap at the driver's max of 50.
-            amount: BigInt(Math.min(50, v1Params.scrollAmount ?? 3)),
+            amount: BigInt(Math.min(50, desktopParams.scrollAmount ?? 3)),
             ...point,
           },
           signal,
@@ -357,49 +357,49 @@ async function handleDesktopAct(
       break;
     }
     default: {
-      const frame = await currentFrame(driver, frameState, v1Params, signal);
-      switch (v1Params.action) {
+      const frame = await currentFrame(driver, frameState, desktopParams, signal);
+      switch (desktopParams.action) {
         case "left_click":
           assertToolSuccess(
-            await driver.click(clickArgs(frame, v1Params, ClickButton.Left, 1), signal),
+            await driver.click(clickArgs(frame, desktopParams, ClickButton.Left, 1), signal),
             "click",
           );
           break;
         case "right_click":
           assertToolSuccess(
-            await driver.click(clickArgs(frame, v1Params, ClickButton.Right, 1), signal),
+            await driver.click(clickArgs(frame, desktopParams, ClickButton.Right, 1), signal),
             "click",
           );
           break;
         case "middle_click":
           assertToolSuccess(
-            await driver.click(clickArgs(frame, v1Params, ClickButton.Middle, 1), signal),
+            await driver.click(clickArgs(frame, desktopParams, ClickButton.Middle, 1), signal),
             "click",
           );
           break;
         case "double_click":
           assertToolSuccess(
-            await driver.click(clickArgs(frame, v1Params, ClickButton.Left, 2), signal),
+            await driver.click(clickArgs(frame, desktopParams, ClickButton.Left, 2), signal),
             "click",
           );
           break;
         case "triple_click":
           assertToolSuccess(
-            await driver.click(clickArgs(frame, v1Params, ClickButton.Left, 3), signal),
+            await driver.click(clickArgs(frame, desktopParams, ClickButton.Left, 3), signal),
             "click",
           );
           break;
         case "mouse_move": {
-          const point = scalePoint(frame, v1Params.x, v1Params.y, v1Params.action);
+          const point = scalePoint(frame, desktopParams.x, desktopParams.y, desktopParams.action);
           assertToolSuccess(await driver.moveCursor(point, signal), "move_cursor");
           break;
         }
         case "left_click_drag": {
-          const from = scalePoint(frame, v1Params.fromX, v1Params.fromY, "drag start");
-          const to = scalePoint(frame, v1Params.x, v1Params.y, "drag end");
+          const from = scalePoint(frame, desktopParams.fromX, desktopParams.fromY, "drag start");
+          const to = scalePoint(frame, desktopParams.x, desktopParams.y, "drag end");
           // The typed desktop drag API has no modifier field. Refuse instead of
           // silently widening a model request into an unmodified drag.
-          if (normalizeModifiers(v1Params.modifiers).length > 0) {
+          if (normalizeModifiers(desktopParams.modifiers).length > 0) {
             throw new Error(
               "COMPUTER_UNSUPPORTED_ACTION: modifier-held drag is unsupported by cua-driver",
             );
@@ -413,9 +413,9 @@ async function handleDesktopAct(
                 toY: to.y,
                 // CUA caps desktop drag duration at 10 seconds; clamp rather than
                 // rejecting a valid computer.act request at the SDK boundary.
-                ...(v1Params.durationMs === undefined
+                ...(desktopParams.durationMs === undefined
                   ? {}
-                  : { durationMs: BigInt(Math.min(10_000, v1Params.durationMs)) }),
+                  : { durationMs: BigInt(Math.min(10_000, desktopParams.durationMs)) }),
               },
               signal,
             ),
@@ -583,7 +583,7 @@ export function createCuaComputerProvider(
                   : "COMPUTER_DRIVER_UNAVAILABLE: cua-computer supports macOS, Windows, and Linux",
               );
             }
-            return await handleV2Act(
+            return await handleWindowAct(
               platform,
               executionDriver,
               frameState,
